minor cleanup, matrix bot get's own file
